# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Comics documentation",
        "description": "V1 Api docs",
        "contact": {
            "email": ""
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/manga": {
            "get": {
                "tags": [
                    "Mangas"
                ],
                "summary": "Список манг",
                "description": "Список манг",
                "operationId": "89f6e81c7206269aebf4f9f9195f6023",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[teams]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter[tags]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MangaResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ]
            }
        },
        "/manga/{slug}": {
            "get": {
                "tags": [
                    "Mangas"
                ],
                "summary": "Страница манги",
                "description": "Страница манги",
                "operationId": "511b53d51e7360d6efcb33e2d31f1d20",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MangaResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found."
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ]
            }
        },
        "/mangas": {
            "post": {
                "tags": [
                    "Mangas"
                ],
                "summary": "Добавить мангу",
                "description": "Добавить мангу",
                "operationId": "c237b7d2aeb9ba03a8e048e940dc7514",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MangaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MangaResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ]
            }
        },
        "/manga/{id}": {
            "delete": {
                "tags": [
                    "Mangas"
                ],
                "summary": "Обновить мангу",
                "description": "Обновить мангу",
                "operationId": "4e0e1d3e392650777333de7060158776",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MangaResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Mangas"
                ],
                "summary": "Обновить мангу",
                "description": "Обновить мангу",
                "operationId": "3ac5b136f41593b970d3b34ce748184e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MangaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MangaResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Manga": {
                "title": "Manga",
                "description": "Manga Model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer"
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string"
                    },
                    "slug": {
                        "title": "slug",
                        "description": "slug",
                        "type": "string"
                    },
                    "description": {
                        "title": "description",
                        "description": "description",
                        "type": "string"
                    },
                    "media": {
                        "title": "media",
                        "description": "media",
                        "type": "string"
                    },
                    "tags": {
                        "title": "tags",
                        "description": "tags",
                        "type": "string"
                    },
                    "rating": {
                        "title": "rating",
                        "description": "rating",
                        "type": "string"
                    },
                    "ratings_count": {
                        "title": "ratings_count",
                        "description": "ratings_count",
                        "type": "string"
                    },
                    "votes": {
                        "title": "votes",
                        "description": "votes",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Manga"
                }
            },
...
            "MangaRequest": {
                "title": "Manga request",
                "description": "Manga request body data",
                "required": [
                    "name",
                    "description",
                    "tags"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string"
                    },
                    "description": {
                        "title": "description",
                        "description": "description",
                        "type": "string"
                    },
                    "published_at": {
                        "title": "published_at",
                        "description": "published_at",
                        "type": "string"
                    },
                    "tags": {
                        "title": "tags",
                        "description": "tags",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "media": {
                        "title": "media",
                        "description": "media",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "MangaResource": {
                "title": "MangaResource",
                "description": "Manga Resource",
                "properties": {
                    "manga": {
                        "$ref": "#/components/schemas/Manga"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Manga Resource"
                }
            },
...
        },
        "securitySchemes": {
            "api-key": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}